[feat] 관광지 상세 정보 response API 구현

+++ src/main/java/academic_festival/gyeonggi_go/Place/Repository/PlaceRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PlaceRepository extends JpaRepository<Place, Integer> {
+public interface PlaceRepository extends JpaRepository<Place, Long> {
     Optional<Place> findByPlaceNameAndAddress(String placeName, String address);
 }
